+++ README.md
@@ -2,3 +2,8 @@ step running
 npm install
 docker-compose build
 docker-compose up
+
+docker-compose build // If you add/upgrade anything better to build
+docker-compose up // Will start all the services
+/* Or you could use */
+docker-compose up --build
