--- src/elasticsearch/auth/dto/refresh-token.dto.ts
@@ -1,8 +1,0 @@
-import { ApiProperty } from "@nestjs/swagger";
-import { IsString } from "class-validator";
-
-export class RefreshTokenDto {
-    @ApiProperty()
-    @IsString()
-    refresh_token: string
-}
