--- src/elasticsearch/auth/interface/jwt-payload.interface.ts
@@ -1,6 +1,0 @@
-import { UserInfo } from "../../user/entity/user-info.entity";
-
-export interface JwtPayload {
-    username: string
-    user_info: UserInfo
-}
