--- src/elasticsearch/auth/jwt-strategy.ts
@@ -1,36 +1,0 @@
-import { Injectable, UnauthorizedException } from "@nestjs/common";
-import { PassportStrategy } from "@nestjs/passport";
-import { InjectRepository } from "@nestjs/typeorm";
-import { Strategy, ExtractJwt } from 'passport-jwt'
-import { User } from "./entity/user.entity";
-import { JwtPayload } from "./interface/jwt-payload.interface";
-import { UserRepository } from "./repository/user.repository";
-
-import * as config from 'config'
-
-const dbConfig = config.get('jwt')
-
-@Injectable()
-export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
-    constructor(
-        @InjectRepository(UserRepository)
-        private userRepository: UserRepository
-    ) {
-        super({
-            jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-            secretOrKey: process.env.JWT_ACCESS_TOKEN_SECRET || dbConfig.secret
-        })
-    }
-
-    async validate(payload: JwtPayload): Promise<User> {
-        const { username } = payload
-        const user = await this.userRepository.findOne({ 
-            where: { username: username }
-         })
-
-        if (!user) {
-            throw new UnauthorizedException()
-        }
-        return user
-    }
-}
