--- src/elasticsearch/guards/jwt-authentication.guard.ts
@@ -1,5 +1,0 @@
-import { Injectable } from "@nestjs/common";
-import { AuthGuard } from "@nestjs/passport";
-
-@Injectable()
-export class JwtAuthenticationGuard extends AuthGuard('jwt') {}
