--- src/elasticsearch/guards/jwt-refresh-token.guard.ts
@@ -1,5 +1,0 @@
-import { Injectable } from "@nestjs/common";
-import { AuthGuard } from "@nestjs/passport";
-
-@Injectable()
-export class JwtRefreshTokenGuard extends AuthGuard('jwt-refresh-token') {}
