--- src/elasticsearch/user/dto/user-info.dto.ts
@@ -1,30 +1,0 @@
-import { ApiProperty } from "@nestjs/swagger"
-
-export class UserInfoDto {
-    @ApiProperty({
-        required: false
-    })
-    petName: string
-
-    @ApiProperty({
-        type: 'file',
-        properties: {
-            file: {
-                type: 'string',
-                format: 'binary'
-            }
-        },
-        required: false
-    })
-    photo: string
-    
-    @ApiProperty({
-        required: false
-    })
-    modified_photo: string
-
-    @ApiProperty({
-        required: false
-    })
-    address: string
-}
