--- src/elasticsearch/user/entity/user-info.entity.ts
@@ -1,19 +1,0 @@
-import { BaseEntity, Column, Entity, PrimaryGeneratedColumn } from "typeorm";
-
-@Entity()
-export class UserInfo extends BaseEntity {
-    @PrimaryGeneratedColumn()
-    id: number
-
-    @Column({ name: 'pet_name', type: "varchar", nullable: true })
-    petName: string
-
-    @Column({ type: "varchar", nullable: true })
-    photo: string
-
-    @Column({ type: "varchar", nullable: true })
-    modified_photo: string
-
-    @Column({ type: "varchar", nullable: true })
-    address: string
-}
