--- src/elasticsearch/user/interface/user-info.interface.ts
@@ -1,8 +1,0 @@
-export interface userInfoData {
-    id: number,
-    username: string,
-    petName?: string,
-    photo?: string,
-    modified_photo?: string,
-    address: string
-}
