--- src/elasticsearch/user/repository/user-info.repository.ts
@@ -1,6 +1,0 @@
-import { EntityRepository, Repository } from "typeorm";
-import { UserInfo } from "../entity/user-info.entity";
-
-@EntityRepository(UserInfo)
-export class UserInfoRepository extends Repository<UserInfo> {
-}
