+++ src/todosearch/dto/create-todosearch.dto.ts
@@ -1,0 +1,1 @@
+export class CreateTodosearchDto {}
