+++ src/todosearch/dto/update-todosearch.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/swagger';
+import { CreateTodosearchDto } from './create-todosearch.dto';
+
+export class UpdateTodosearchDto extends PartialType(CreateTodosearchDto) {}
