+++ src/todosearch/types/todoSearchResult.interface.ts
