+++ src/users/decorators/get-user.decorator.ts
@@ -1,0 +1,8 @@
+import { createParamDecorator, ExecutionContext } from "@nestjs/common";
+import User from "../user.entity";
+
+
+export const GetUser = createParamDecorator((data, ctx: ExecutionContext): User => {
+    const req = ctx.switchToHttp().getRequest();
+    return req.user
+})
